+++ test/videotilecontroller/DefaultVideoTileController.test.ts
@@ -117,6 +117,25 @@ describe('DefaultVideoTileController', () => {
       tileController.bindVideoElement(tileId, videoElement);
       tileController.unbindVideoElement(tileId);
     });
+
+    it('clears srcObject with unbinding a video element', () => {
+      let videoElement = document.createElement('video');
+      const tileId = tileController.addVideoTile().id();
+      tileController.bindVideoElement(tileId, videoElement);
+      const tile = tileController.getVideoTile(tileId);
+      tile.bindVideoStream('attendee', false, mockMediaStream, 1, 1, 1);
+      expect(videoElement.srcObject).to.eq(mockMediaStream);
+      tileController.unbindVideoElement(tileId);
+      expect(videoElement.srcObject).to.eq(null);
+      videoElement = null;
+    });
+
+    it('ignores if no tile bound to a tileId', () => {
+      const loggerSpy = sinon.spy(audioVideoController.logger, 'warn');
+      tileController.unbindVideoElement(0);
+      expect(loggerSpy.calledWith('Ignoring video element un-binding for unknown tile id 0')).to.be
+        .true;
+    });
   });
 
   describe('startLocalVideoTile', () => {